refactor(tests): split ha node-state coverage module
