+++ src/server/tests.rs
@@ -33,6 +33,7 @@ mod tests {
     mod admin_users_and_tokens;
     mod alerts_and_ha;
     mod alerts_and_ha_event_exports;
+    mod alerts_and_ha_node_state;
     mod alerts_and_ha_serving_modes;
     mod alerts_and_ha_startup_roles;
     mod api_keys_and_registration;
